requirements.txt
flask==3.1.2
gunicorn==23.0.0
tensorflow-cpu==2.15.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
